Render configuration as MetalLB CRs
We implement logic which transforms existing configurations into
CR-based configurations.

The commit also updates the default test case to use the new config
parameters which should used to generate CR-based configurations and
adds a `legacy` test case which configures MetalLB through the
deprecated `config` parameter.

The transformation logic has been verified by manually comparing the
results of the `legacy` test case with the output of the official
configuration conversion tool.

The CRD-based configuration is mandatory for MetalLB 0.13.2 and newer.